+++ src/main/java/net/pms/store/StoreItem.java
@@ -528,21 +528,15 @@ public abstract class StoreItem extends StoreResource {
 					isIncompatible = true;
 					LOGGER.debug(prependTranscodingReason + "the bitrate ({} b/s) is too high ({} b/s).", getName(), mediaInfo.getBitRate(),
 							maxBandwidth);
-				} else if (renderer.isH264Level41Limited() && mediaVideo.isH264()) {
+				} else if (mediaVideo.isH264()) {
+					double h264LevelLimit = renderer.getH264LevelLimit();
 					if (mediaVideo.getFormatLevel() != null) {
-						double h264Level = 4.1;
-
-						try {
-							h264Level = Double.parseDouble(mediaVideo.getFormatLevel());
-						} catch (NumberFormatException e) {
-							LOGGER.trace("Could not convert {} to double: {}", mediaVideo.getFormatLevel(), e.getMessage());
-						}
-
-						if (h264Level > 4.1) {
+						double h264Level = mediaVideo.getFormatLevelAsDouble(4.1);
+						if (h264Level > h264LevelLimit) {
 							isIncompatible = true;
-							LOGGER.debug(prependTranscodingReason + "the H.264 level ({}) is not supported.", getName(), h264Level);
+							LOGGER.debug(prependTranscodingReason + "the H.264 level ({}) is not supported by the renderer (limit: {}).", getName(), h264Level, h264LevelLimit);
 						}
-					} else {
+					} else if (h264LevelLimit < 4.2) {
 						isIncompatible = true;
 						LOGGER.debug(prependTranscodingReason + "the H.264 level is unknown.", getName());
 					}
@@ -583,10 +577,6 @@ public abstract class StoreItem extends StoreResource {
 	 * @return String representation of the mime type
 	 */
 	public String getRendererMimeType() {
-		// FIXME: There is a flaw here. In addChild(StoreResource) the mime type
-		// is determined for the default renderer. This renderer may rewrite the
-		// mime type based on its configuration. Looking up that mime type is
-		// not guaranteed to return a match for another renderer.
 		String mime = renderer.getMimeType(this);
 
 		// Use our best guess if we have no valid mime type